+++ src/web/src/app/globals.css
@@ -64,6 +64,8 @@
     --border: 0 0% 14.9%;
     --input: 0 0% 14.9%;
     --ring: 0 0% 83.1%;
+
+    --shadow-1: 0 88.2% 50.2%;
   }
 }
 
@@ -78,5 +80,41 @@
   html {
     @apply antialiased;
     scroll-behavior: smooth;
+  }
+}
+
+.shadow-anim {
+  position: absolute;
+  background: hsl(var(--shadow-1) / 0%);
+  border-radius: 24px;
+  rotate: 35deg;
+
+  width: 200px;
+  top: 170px;
+  height: 350px;
+  filter: blur(150px);
+  animation: shadow-slide infinite 4s linear alternate;
+}
+
+@keyframes shadow-slide {
+  from {
+    background: hsl(var(--shadow-1) / 20%);
+    right: 260px;
+  }
+  to {
+    background: hsl(var(--shadow-1) / 80%);
+    right: 160px;
+  }
+}
+
+@media (max-width: 1024px) {
+  .shadow-anim {
+    display: none;
+  }
+}
+
+@media (max-width: 768px) {
+  .shadow-anim {
+    display: none;
   }
 }
