+++ src/web/src/app/layout.tsx
@@ -6,8 +6,8 @@ import "./globals.css";
 const inter = Inter({ subsets: ["latin"] });
 
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "NextJudge",
+  description: "A Next Generation Platform for Competitive Programming.",
 };
 
 export default function RootLayout({
