+++ src/web/src/app/page.tsx
@@ -5,7 +5,7 @@ export default function Home() {
   return (
     <>
       <Navbar />
-      <main className="flex min-h-screen flex-col items-center justify-between overflow-x-hidden">
+      <main className="flex max-w-screen flex-col items-center justify-between overflow-x-hidden">
         <Hero />
       </main>
     </>
