+++ src/web/src/components/landing/hero.tsx
@@ -1,6 +1,6 @@
+import { Button, buttonVariants } from "@/components/ui/button";
 import { GitHubLogoIcon } from "@radix-ui/react-icons";
 import Image from "next/image";
-import { Button, buttonVariants } from "../ui/button";
 export const Hero = () => {
   return (
     <section className="container grid lg:grid-cols-2 place-items-center py-20 md:py-32 gap-10">
@@ -8,9 +8,9 @@ export const Hero = () => {
         <main className="text-5xl md:text-6xl font-bold">
           <h1 className="inline">
             <span className="inline bg-gradient-to-r from-osu  to-osubrown text-transparent bg-clip-text">
-              Ergonomic
+              Next Generation
             </span>{" "}
-            Code Judge
+            Platform
           </h1>{" "}
           for{" "}
           <h2 className="inline">
@@ -22,9 +22,9 @@ export const Hero = () => {
         </main>
 
         <p className="text-xl text-muted-foreground md:w-10/12 mx-auto lg:mx-0">
-          NextJudge is a modern, ergonomic and feature-rich code judge for
-          competitive programming. It is designed to be blazing fast, reliable
-          and user-friendly.
+          NextJudge is a modern, open-source platform for hosting and partaking
+          in competitive programming. It is designed to be easy to use, fast,
+          and extensible.
         </p>
 
         <div className="space-y-4 md:space-y-0 md:space-x-4">
@@ -45,15 +45,16 @@ export const Hero = () => {
       </div>
 
       {/* Hero preview */}
-      <div className="z-10">
+      <div className="z-10 lg:scale-125">
         <Image
           src="/preview.png"
           alt="Editor Preview"
           width={1000}
           height={1000}
-          className="w-[48rem] dark:shadow-osu max-w-none rounded-lg ring-2 ring-orange-500 dark:ring-orange-600/20 ring-opacity-10 border-8 border-white border-opacity-5 sm:w-[50rem] md:-ml-4 lg:-ml-0"
+          className="rounded-lg ring-1 ring-osu/20 bg-black w-fit h-fit lg:w-full lg:h-full"
         />
       </div>
+      <div className="shadow-anim"></div>
     </section>
   );
 };
