+++ src/web/src/components/navbar.tsx
@@ -17,6 +17,8 @@ import {
 } from "@/components/ui/sheet";
 import { useState } from "react";
 
+import { Button } from "@/components/ui/button";
+import { directoryRoutes, platformRoutes, routeList } from "@/lib/constants";
 import { cn } from "@/lib/utils";
 import { GitHubLogoIcon } from "@radix-ui/react-icons";
 import { Menu, Pyramid } from "lucide-react";
@@ -24,34 +26,10 @@ import { Icons } from "./icons";
 import { ModeToggle } from "./theme";
 import { buttonVariants } from "./ui/button";
 
-interface RouteProps {
-  href: string;
-  label: string;
-}
-
-export const routeList: RouteProps[] = [
-  {
-    href: "/platform",
-    label: "Home",
-  },
-  {
-    href: "/#features",
-    label: "Features",
-  },
-  {
-    href: "/auth/signup",
-    label: "Sign Up",
-  },
-  {
-    href: "/auth/login",
-    label: "Login",
-  },
-];
+// having three nav components in this file is kinda confusing we'll need to refactor this later
 
 export default function Navbar() {
   const [isOpen, setIsOpen] = useState<boolean>(false);
-  // useState here explicitly lets Next.js know this is a 'client' component
-  const [isClient] = useState<boolean>(false);
   return (
     <header className="sticky border-b-[1px] top-0 z-40 w-full bg-white dark:border-b-neutral-500/40 dark:bg-background">
       <NavigationMenu className="mx-auto">
@@ -64,55 +42,51 @@ export default function Navbar() {
           </NavigationMenuItem>
 
           {/* mobile */}
-          {isClient && (
-            <div className="flex md:hidden">
-              <ModeToggle />
-              <Sheet open={isOpen} onOpenChange={setIsOpen}>
-                <SheetTrigger className="px-2">
-                  <Menu
-                    className="flex md:hidden h-5 w-5"
-                    onClick={() => setIsOpen(true)}
-                  >
-                    <span className="sr-only">Menu Icon</span>
-                  </Menu>
-                </SheetTrigger>
 
-                <SheetContent side={"left"}>
-                  <SheetHeader>
-                    <SheetTitle className="font-bold text-xl">
-                      NextJudge
-                    </SheetTitle>
-                  </SheetHeader>
-                  <nav className="flex flex-col justify-center items-center gap-2 mt-4">
-                    {routeList.map(({ href, label }: RouteProps) => (
-                      <a
-                        key={label}
-                        href={href}
-                        onClick={() => setIsOpen(false)}
-                        className={buttonVariants({ variant: "ghost" })}
-                      >
-                        {label}
-                      </a>
-                    ))}
+          <div className="flex md:hidden">
+            <ModeToggle />
+            <Sheet open={isOpen} onOpenChange={setIsOpen}>
+              <SheetTrigger className="px-2" asChild>
+                <Button variant="ghost">
+                  <Menu className="h-5 w-5" onClick={() => setIsOpen(true)} />
+                </Button>
+              </SheetTrigger>
+
+              <SheetContent side={"left"}>
+                <SheetHeader>
+                  <SheetTitle className="font-bold text-xl">
+                    NextJudge
+                  </SheetTitle>
+                </SheetHeader>
+                <nav className="flex flex-col justify-center items-center gap-2 mt-4">
+                  {routeList.map(({ href, label }) => (
                     <a
-                      href="https://github.com/nextjudge/nextjudge"
-                      target="_blank"
-                      className={`w-[110px] border ${buttonVariants({
-                        variant: "secondary",
-                      })}`}
+                      key={label}
+                      href={href}
+                      onClick={() => setIsOpen(false)}
+                      className={buttonVariants({ variant: "ghost" })}
                     >
-                      <GitHubLogoIcon className="mr-2 w-5 h-5" />
-                      Github
+                      {label}
                     </a>
-                  </nav>
-                </SheetContent>
-              </Sheet>
-            </div>
-          )}
+                  ))}
+                  <a
+                    href="https://github.com/nextjudge/nextjudge"
+                    target="_blank"
+                    className={`w-[110px] border ${buttonVariants({
+                      variant: "secondary",
+                    })}`}
+                  >
+                    <GitHubLogoIcon className="mr-2 w-5 h-5" />
+                    Github
+                  </a>
+                </nav>
+              </SheetContent>
+            </Sheet>
+          </div>
 
           {/* desktop */}
           <nav className="hidden md:flex gap-4">
-            {routeList.map((route: RouteProps, i) => (
+            {routeList.map((route, i) => (
               <a
                 href={route.href}
                 key={i}
@@ -131,55 +105,11 @@ export default function Navbar() {
   );
 }
 
-const platformRoutes: RouteProps[] = [
-  {
-    href: "/platform",
-    label: "Home",
-  },
-  {
-    href: "/platform/contests",
-    label: "Contests",
-  },
-  {
-    href: "/platform/problems",
-    label: "Problems",
-  },
-  {
-    href: "/platform/admin",
-    label: "Admin",
-  },
-];
-
-const directoryRoutes = {
-  infosNav: [
-    {
-      title: "Directory",
-      items: [
-        {
-          title: "Recent Submissions",
-          href: "/platform/problems#submissions",
-          description: "Tried submitting a solution? Here's your latest.",
-        },
-        {
-          title: "Upcoming Contests",
-          href: "/platform/contests",
-          description: "Check out the upcoming contests.",
-        },
-        {
-          title: "Editorials",
-          href: "/platform/editorials",
-          description: "Read the editorials for the problems.",
-        },
-      ],
-    },
-  ],
-};
-
 export function MainNavigationMenu() {
   const infos = directoryRoutes.infosNav[0];
   return (
     <NavigationMenu>
-      <NavigationMenuList className="flex gap-4 max-w-screen">
+      <NavigationMenuList className="flex gap-0 max-w-screen">
         <NavigationMenuItem>
           <NavigationMenuTrigger>{infos.title}</NavigationMenuTrigger>
           <NavigationMenuContent>
@@ -187,7 +117,7 @@ export function MainNavigationMenu() {
               <li className="row-span-3">
                 <a
                   className="flex h-full w-full select-none flex-col justify-end rounded-md bg-gradient-to-b from-muted/50 to-muted p-6 no-underline outline-none focus:shadow-md"
-                  href="/platform/problems/1"
+                  href="/platform/problems/problem/1"
                 >
                   <Pyramid className="w-10 h-10" />
                   <div className="mb-2 mt-3 text-lg font-medium text-gradient_blaze-orange">
@@ -204,18 +134,16 @@ export function MainNavigationMenu() {
             </ul>
           </NavigationMenuContent>
         </NavigationMenuItem>
-
-        {platformRoutes.map((link) => (
-          <NavigationMenuItem>
-            <a
-              key={link.href}
-              href={link.href}
-              className={navigationMenuTriggerStyle()}
-            >
-              {link.label}
-            </a>
-          </NavigationMenuItem>
-        ))}
+        {/* holy hack */}
+        <li className="flex">
+          {platformRoutes.map((link) => (
+            <NavigationMenuItem key={link.label}>
+              <a href={link.href} className={navigationMenuTriggerStyle()}>
+                {link.label}
+              </a>
+            </NavigationMenuItem>
+          ))}
+        </li>
       </NavigationMenuList>
     </NavigationMenu>
   );
@@ -284,7 +212,7 @@ export function PlatformNavbar() {
                 <SheetTitle className="font-bold text-xl">NextJudge</SheetTitle>
               </SheetHeader>
               <nav className="flex flex-col justify-center items-center gap-2 mt-4">
-                {routeList.map(({ href, label }: RouteProps) => (
+                {routeList.map(({ href, label }) => (
                   <a
                     key={label}
                     href={href}
