+++ src/web/src/lib/constants.ts
@@ -34,3 +34,71 @@ export const links = [
     ],
   },
 ];
+
+export const directoryRoutes = {
+  infosNav: [
+    {
+      title: "Directory",
+      items: [
+        {
+          title: "Recent Submissions",
+          href: "/platform/problems#submissions",
+          description: "Tried submitting a solution? Here's your latest.",
+        },
+        {
+          title: "Upcoming Contests",
+          href: "/platform/contests",
+          description: "Check out the upcoming contests.",
+        },
+        {
+          title: "Editorials",
+          href: "/platform/editorials",
+          description: "Read the editorials for the problems.",
+        },
+      ],
+    },
+  ],
+};
+
+interface RouteProps {
+  href: string;
+  label: string;
+}
+
+export const routeList: RouteProps[] = [
+  {
+    href: "/platform",
+    label: "Home",
+  },
+  {
+    href: "/#features",
+    label: "Features",
+  },
+  {
+    href: "/auth/signup",
+    label: "Sign Up",
+  },
+  {
+    href: "/auth/login",
+    label: "Login",
+  },
+];
+
+export const platformRoutes: RouteProps[] = [
+  {
+    href: "/platform",
+    label: "Home",
+  },
+  {
+    href: "/platform/contests",
+    label: "Contests",
+  },
+  {
+    href: "/platform/problems",
+    label: "Problems",
+  },
+  {
+    href: "/platform/admin",
+    label: "Admin",
+  },
+];
